accessible/tests/mochitest/relations/test_shadowdom.html

<html>

<head>
  <title>Explicit content and shadow DOM content relations tests</title>
  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../relations.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>

  <script type="application/javascript">
    function doTest() {
      let iframeDoc = document.getElementById("iframe").contentDocument;

      // explicit content
      let label = iframeDoc.getElementById("label");
      let element = iframeDoc.getElementById("element");
      testRelation(label, RELATION_LABEL_FOR, element);
      testRelation(element, RELATION_LABELLED_BY, label);

      // shadow DOM content
      let shadowRoot = iframeDoc.getElementById("shadowcontainer").shadowRoot;
      let shadowLabel = shadowRoot.getElementById("label");
      let shadowElement = shadowRoot.getElementById("element");

      testRelation(shadowLabel, RELATION_LABEL_FOR, shadowElement);
      testRelation(shadowElement, RELATION_LABELLED_BY, shadowLabel);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    SpecialPowers.pushPrefEnv({
      set: [
        [ "dom.webcomponents.shadowdom.enabled", true ],
      ],
    }, function() {
      // This test loads in an iframe, to ensure that the element instance is
      // loaded with the correct value of the preference.
      let sc = "script";
      let iframe = document.createElement("iframe");
      iframe.id = "iframe";
      iframe.src = `data:text/html,<html>
        <body>
          <div id='label'></div><div id='element' aria-labelledby='label'></div>
          <div id='shadowcontainer'></div>
          <${sc}>
            let shadowRoot = document.getElementById('shadowcontainer').
              attachShadow({mode: 'open'});
            shadowRoot.innerHTML =
              "<div id='label'></div><div id='element' aria-labelledby='label'></div>";
          </${sc}>
        </body>
      </html>`;

      addA11yLoadEvent(doTest, iframe.contentWindow);
      document.body.appendChild(iframe);
    });
  </script>
</head>

<body>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>
</body>
</html>
